old_tests/coap-tests-dev-tests_/coap-tests-dev-tests/requirements.txt
cffi==1.16.0
CoAPthon3==1.0.2
cryptography==42.0.5
pycparser==2.22
python3-dtls==1.3.0
